Compute energies and normalize pushed potential ground state energies

+++ writing.h
@@ -11,7 +11,7 @@ void fwrite_vector_complex_arg(FILE *f, const gsl_vector_complex *psi);
 void fwrite_matrix_complex(FILE *f, const gsl_matrix_complex *M);
 void fwrite_matrix(FILE *f, const gsl_matrix *M);
 
-void downsample_vector_complex_alloc(gsl_vector_complex **vv, const gsl_vector_complex *v, unsigned int n);
+void downsample_vector_complex_alloc(gsl_vector_complex **vv, const gsl_vector_complex *v, size_t n);
 
 void terminal_graph_abs(const gsl_vector_complex *psi, const int, const double);
 void terminal_graph_abs2(const gsl_vector_complex *psi, const int, const double);
